// src/api/generated/models/monitoringTableResult.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface MonitoringTableResult {
  queryId?: string;
  resource?: string;
  status?: string;
  pageNum?: number;
  pageSize?: number;
  total?: string;
  records?: unknown[];
  errorMessage?: string;
}